fix disposing for multipart and single part + unittests for singlepart
